Added USB Captures
Captures directory contains a number of activities being captured.

typing1 is me typing a small sentence, mostly to test that the
keyboard functionality still works.

getfwversion(S1946V15) is using the stock tool to get the firmware
version off the keyboard, the result of which was S1946V15.

the flashing directory contains 2 instances of using the stock
tool to flash firmware onto the keyboard.  original is the file
"Poker II original firmware.bin".  nkro is
"Poker II NKRO in USB mode.bin".